Update README.md
In this project i decided to do a normal Simon game. It is a memory game where you have to follow the sequence and copy the sequence correctly to move on to the next level. The mode is pretty straight forward. When a coloured button flashes, you click on the button. When you get further on in the game you will have several colours to remember in order. The game gets harder the longer you play! 

Technologies Used

jQuery - I have used jQuery to simplify DOM manipulation - I have used CSS and favicon for effects Font Awesome and icons respectively- I have used font awesome to add some different font onto my website. CSS - For styling the game  & HTML - For the elements of the game.